room regist facility (#36)

+++ src/main/java/com/yts8/sixuniverse/member/repository/MemberMapper.java
@@ -8,7 +8,7 @@ public interface MemberMapper {
 
   MemberDto findByEmail(String email);
 
-  MemberDto findById(int memberId);
+  MemberDto findById(Long memberId);
 
   void save(MemberDto memberDto);
